Change version tag to include v
As discussed in imagemin/gifsicle-bin#55, some imagemin bin packages prepend their version
tags with v and some don't. This PR changes the behavior of this package to prepend tags 
with`v`, as do the majority of packages. This reduces the probability that a wrong tag 
is published by mistake.

Merging this PR means that the next tag must be prepended with a v!